requirements.txt
# The libraries install for this project

flask==3.0.0
flask-cors==4.0.0
pandas==2.1.4
numpy==1.26.2
scikit-learn==1.3.2
matplotlib==3.8.2
seaborn==0.13.0
python-dotenv==1.0.0
openai==1.6.1
google-generativeai==0.3.2
groq==0.9.0
mysql-connector-python==8.2.0